Add missing security restrictions (!492)

+++ library/src/main/kotlin/eu/tib/orkg/prototype/statements/services/StatementService.kt
@@ -1,5 +1,6 @@
 package eu.tib.orkg.prototype.statements.services
 
+import eu.tib.orkg.prototype.auth.domain.UserService
 import eu.tib.orkg.prototype.contributions.domain.model.ContributorId
 import eu.tib.orkg.prototype.spring.spi.FeatureFlagService
 import eu.tib.orkg.prototype.statements.api.BundleConfiguration
@@ -22,6 +23,7 @@ import eu.tib.orkg.prototype.statements.domain.model.StatementId
 import eu.tib.orkg.prototype.statements.domain.model.StatementRepresentation
 import eu.tib.orkg.prototype.statements.domain.model.Thing
 import eu.tib.orkg.prototype.statements.domain.model.ThingId
+import eu.tib.orkg.prototype.statements.spi.OwnershipInfo
 import eu.tib.orkg.prototype.statements.spi.PredicateRepository
 import eu.tib.orkg.prototype.statements.spi.StatementRepository
 import eu.tib.orkg.prototype.statements.spi.TemplateRepository
@@ -42,7 +44,8 @@ class StatementService(
     private val predicateService: PredicateRepository,
     private val statementRepository: StatementRepository,
     private val templateRepository: TemplateRepository,
-    private val flags: FeatureFlagService
+    private val flags: FeatureFlagService,
+    private val userService: UserService,
 ) : StatementUseCases {
 
     override fun findAll(pagination: Pageable): Iterable<StatementRepresentation> =
@@ -139,11 +142,35 @@ class StatementService(
 
     override fun totalNumberOfStatements(): Long = statementRepository.count()
 
-    override fun remove(statementId: StatementId) =
-        statementRepository.deleteByStatementId(statementId)
+    /**
+     * Delete a statement if it is owned by the given contributor, or if the contributor has curator status.
+     *
+     * @param statementId the ID of the statement to delete.
+     * @param contributorId the ID of the contributor requesting the deletion.
+     */
+    override fun delete(statementId: StatementId, contributorId: ContributorId) {
+        val statement = statementRepository.findByStatementId(statementId).orElse(null) ?: return
+        val contributor = userService.findById(contributorId.value).orElse(null)
+        if (statement.isOwnedBy(contributorId) || contributor?.isCurator == true) {
+            statementRepository.deleteByStatementId(statementId)
+        }
+    }
 
-    override fun remove(statementIds: Set<StatementId>) =
-        statementRepository.deleteByStatementIds(statementIds)
+    /**
+     * Delete a set of statements if they are all owned by the given contributor, or if the contributor has curator
+     * status.
+     *
+     * @param statementIds the set of IDs of the statements to delete.
+     * @param contributorId the ID of the contributor requesting the deletion.
+     */
+    override fun delete(statementIds: Set<StatementId>, contributorId: ContributorId) {
+        val ownerInfo = statementRepository.determineOwnership(statementIds)
+        val allOwnedByContributor = ownerInfo.map(OwnershipInfo::owner).all { it == contributorId }
+        val contributor = userService.findById(contributorId.value).orElse(null)
+        if (allOwnedByContributor || contributor?.isCurator == true) {
+            statementRepository.deleteByStatementIds(statementIds)
+        }
+    }
 
     override fun update(command: UpdateStatementUseCase.UpdateCommand) {
         var found = statementRepository.findByStatementId(command.statementId)
